+++ packages/get-starknet/.eslintrc.js
@@ -1,15 +1,42 @@
 module.exports = {
-  env: {
-    browser: true,
-    es2021: true,
-  },
-  extends: ['plugin:@typescript-eslint/recommended'],
-  parser: '@typescript-eslint/parser',
+  extends: ['../../.eslintrc.js'],
+
   parserOptions: {
-    ecmaVersion: 'latest',
-    sourceType: 'module',
+    tsconfigRootDir: __dirname,
   },
-  plugins: ['@typescript-eslint'],
-  rules: {
-  },
+
+  overrides: [
+    {
+      files: ['snap.config.ts'],
+      extends: ['@metamask/eslint-config-nodejs'],
+    },
+    {
+      files: ['*.ts', '*.tsx'],
+      extends: ['@metamask/eslint-config-typescript'],
+      rules: {
+        'valid-jsdoc': 'off',
+        'jsdoc/require-description': 'off',
+        'jsdoc/require-returns': 'off',
+        'jsdoc/require-param-description': 'off',
+        'jsdoc/match-description': 'off',
+      },
+    },
+    {
+      files: ['*.test.ts'],
+      extends: ['@metamask/eslint-config-jest'],
+      rules: {
+        '@typescript-eslint/no-shadow': ['error', { allow: ['describe', 'expect', 'it'] }],
+        '@typescript-eslint/unbound-method': 'off',
+      },
+    },
+
+    {
+      files: ['*.ts'],
+      rules: {
+        'import/no-nodejs-modules': 'off',
+      },
+    },
+  ],
+
+  ignorePatterns: ['!.eslintrc.js', 'dist/', '**/test', '.nyc_output/', 'coverage/'],
 };