+++ packages/get-starknet/webpack.common.js
@@ -1,7 +1,6 @@
 // eslint-disable-next-line @typescript-eslint/no-var-requires
 const path = require('path');
 
-
 module.exports = {
   entry: './src/index.ts',
   output: {