+++ packages/wallet-ui/src/App.tsx
@@ -20,19 +20,28 @@ import { NoMetamaskModal } from 'components/ui/organism/NoMetamaskModal';
 import { MinVersionModal } from './components/ui/organism/MinVersionModal';
 import { useHasMetamask } from 'hooks/useHasMetamask';
 import { DUMMY_ADDRESS } from 'utils/constants';
+import { DeployModal } from 'components/ui/organism/DeployModal';
 
 library.add(fas, far);
 
 function App() {
   const { initSnap, getWalletData, checkConnection } = useStarkNetSnap();
-  const { connected, forceReconnect, provider } = useAppSelector((state) => state.wallet);
-  const { infoModalVisible, minVersionModalVisible, upgradeModalVisible } = useAppSelector((state) => state.modals);
+  const { connected, forceReconnect, provider } = useAppSelector(
+    (state) => state.wallet,
+  );
+  const {
+    infoModalVisible,
+    minVersionModalVisible,
+    upgradeModalVisible,
+    deployModalVisible,
+  } = useAppSelector((state) => state.modals);
   const { loader } = useAppSelector((state) => state.UI);
   const networks = useAppSelector((state) => state.networks);
   const { accounts } = useAppSelector((state) => state.wallet);
   const { hasMetamask } = useHasMetamask();
 
-  const address = accounts?.length > 0 ? (accounts[0] as unknown as string) : DUMMY_ADDRESS;
+  const address =
+    accounts?.length > 0 ? (accounts[0] as unknown as string) : DUMMY_ADDRESS;
 
   useEffect(() => {
     if (!provider) {
@@ -56,7 +65,6 @@ function App() {
   }, [networks.activeNetwork, provider]);
 
   const loading = loader.isLoading;
-
   return (
     <ThemeProvider theme={theme}>
       <GlobalStyle />
@@ -67,17 +75,33 @@ function App() {
         <PopIn isOpen={minVersionModalVisible} showClose={false}>
           <MinVersionModal />
         </PopIn>
-        <PopIn isOpen={!loading && !!hasMetamask && !connected} showClose={false}>
+        <PopIn
+          isOpen={!loading && !!hasMetamask && !connected}
+          showClose={false}
+        >
           <ConnectModal />
         </PopIn>
         <PopIn isOpen={infoModalVisible} showClose={false}>
           <ConnectInfoModal address={address} />
         </PopIn>
-        <PopIn isOpen={!minVersionModalVisible && upgradeModalVisible} showClose={false}>
+        <PopIn
+          isOpen={!minVersionModalVisible && upgradeModalVisible}
+          showClose={false}
+        >
           <UpgradeModel address={address} />
         </PopIn>
+        <PopIn
+          isOpen={!minVersionModalVisible && deployModalVisible}
+          showClose={false}
+        >
+          <DeployModal address={address} />
+        </PopIn>
         <Home address={address} />
-        <PopIn isOpen={loading}>{loading && <LoadingBackdrop>{loader.loadingMessage}</LoadingBackdrop>}</PopIn>
+        <PopIn isOpen={loading}>
+          {loading && (
+            <LoadingBackdrop>{loader.loadingMessage}</LoadingBackdrop>
+          )}
+        </PopIn>
       </FrameworkView>
     </ThemeProvider>
   );