+++ packages/wallet-ui/src/components/ui/organism/SideBar/SideBar.stories.tsx
@@ -6,7 +6,8 @@ export default {
   component: SideBarView,
 } as Meta;
 
-const address = '0x683ec5da50476f84a5d47e822cd4dd35ae3a63c6c1f0725bf28526290d1ee13';
+const address =
+  '0x683ec5da50476f84a5d47e822cd4dd35ae3a63c6c1f0725bf28526290d1ee13';
 
 export const Default = () => {
   return (