+++ packages/wallet-ui/src/types/index.ts
@@ -1,8 +1,14 @@
 import * as Types from '@consensys/starknet-snap/src/types/snapState';
 import { BigNumber } from 'ethers';
 
-export type Account = Pick<Types.AccContract, 'address' | 'publicKey' | 'upgradeRequired'>;
-export type Network = Pick<Types.Network, 'name' | 'chainId' | 'baseUrl' | 'nodeUrl'>;
+export type Account = Pick<
+  Types.AccContract,
+  'address' | 'publicKey' | 'upgradeRequired' | 'deployRequired'
+>;
+export type Network = Pick<
+  Types.Network,
+  'name' | 'chainId' | 'baseUrl' | 'nodeUrl'
+>;
 
 export interface Erc20TokenBalance extends Types.Erc20Token {
   amount: BigNumber;
@@ -31,4 +37,7 @@ export enum TransactionStatus { // for retrieving txn from Starknet feeder gatew
   REJECTED = 'REJECTED',
 }
 
-export type { Erc20Token, Transaction } from '@consensys/starknet-snap/src/types/snapState';
+export type {
+  Erc20Token,
+  Transaction,
+} from '@consensys/starknet-snap/src/types/snapState';